requirements.txt
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
typing_extensions==4.12
psycopg2==2.9.10
django-environ==0.11.2
python-dotenv==1.0.1
